+++ integrations/appboy/lib/index.js
@@ -203,6 +203,7 @@ Appboy.prototype.initializeV2 = function(customEndpoint) {
     openNewsFeedCardsInNewTab: options.openNewsFeedCardsInNewTab,
     requireExplicitInAppMessageDismissal:
       options.requireExplicitInAppMessageDismissal,
+    serviceWorkerLocation: options.serviceWorkerLocation,
     sessionTimeoutInSeconds: Number(options.sessionTimeoutInSeconds) || 30
   };
 
@@ -255,13 +256,31 @@ Appboy.prototype.identify = function(identify) {
   var phone = identify.phone();
   var traits = clone(identify.traits());
 
-  window.appboy.changeUser(userId);
-  window.appboy.getUser().setAvatarImageUrl(avatar);
-  window.appboy.getUser().setEmail(email);
-  window.appboy.getUser().setFirstName(firstName);
-  window.appboy.getUser().setGender(getGender(gender));
-  window.appboy.getUser().setLastName(lastName);
-  window.appboy.getUser().setPhoneNumber(phone);
+  if (userId) {
+    window.appboy.changeUser(userId);
+  }
+  if (avatar) {
+    window.appboy.getUser().setAvatarImageUrl(avatar);
+  }
+  if (email) {
+    window.appboy.getUser().setEmail(email);
+  }
+  if (firstName) {
+    window.appboy.getUser().setFirstName(firstName);
+  }
+  if (gender) {
+    window.appboy.getUser().setGender(getGender(gender));
+  }
+  if (lastName) {
+    window.appboy.getUser().setLastName(lastName);
+  }
+  if (phone) {
+    window.appboy.getUser().setPhoneNumber(phone);
+  }
+  if (address) {
+    window.appboy.getUser().setCountry(address.country);
+    window.appboy.getUser().setHomeCity(address.city);
+  }
   if (address) {
     window.appboy.getUser().setCountry(address.country);
     window.appboy.getUser().setHomeCity(address.city);
@@ -310,10 +329,7 @@ Appboy.prototype.identify = function(identify) {
   // Remove nested hash objects as Braze only supports nested array objects in identify calls
   // https://segment.com/docs/destinations/braze/#identify
   each(function(value, key) {
-    if (
-      typeof value === 'object' &&
-      Array.isArray(value)
-    ) {
+    if (value !== null && typeof value === 'object' && !Array.isArray(value)) {
       delete traits[key];
     }
   }, traits);
@@ -336,7 +352,9 @@ Appboy.prototype.group = function(group) {
   var userId = group.userId();
   var groupIdKey = 'ab_segment_group_' + group.groupId();
 
-  window.appboy.changeUser(userId);
+  if (userId) {
+    window.appboy.changeUser(userId);
+  }
   window.appboy.getUser().setCustomUserAttribute(groupIdKey, true);
 };
 
@@ -367,7 +385,7 @@ Appboy.prototype.track = function(track) {
     delete properties[key];
   }, reserved);
 
-  // Remove nested objects as Braze doesn't support nested objects in tracking calls
+  // Remove nested objects as Braze doesn't support objects in tracking calls
   // https://segment.com/docs/destinations/braze/#track
   each(function(value, key) {
     if (value != null && typeof value === 'object') {
@@ -375,7 +393,9 @@ Appboy.prototype.track = function(track) {
     }
   }, properties);
 
-  window.appboy.changeUser(userId);
+  if (userId) {
+    window.appboy.changeUser(userId);
+  }
   window.appboy.logCustomEvent(eventName, properties);
 };
 
@@ -398,7 +418,9 @@ Appboy.prototype.page = function(page) {
   var eventName = pageEvent.event();
   var properties = page.properties();
 
-  window.appboy.changeUser(userId);
+  if (userId) {
+    window.appboy.changeUser(userId);
+  }
   window.appboy.logCustomEvent(eventName, properties);
 };
 
@@ -419,7 +441,9 @@ Appboy.prototype.orderCompleted = function(track) {
   var currencyCode = track.currency();
   var purchaseProperties = track.properties();
 
-  window.appboy.changeUser(userId);
+  if (userId) {
+    window.appboy.changeUser(userId);
+  }
 
   // remove reduntant properties
   del(purchaseProperties, 'products');